# openapi.yaml
openapi: 3.0.3
info:
  title: Alimenti API
  description: >
    API REST per la gestione di alimenti (composizione CREA/INRAN), tabelle
    di riferimento LARN, ricette, pazienti con piani alimentari e
    appuntamenti, piramide Smartfood. Trasformazione in API REST
    dell'applicativo "Registro Alimenti" (Project Work — Sviluppo di
    un'applicazione full-stack API-based per un'organizzazione del settore
    sanitario, CdS Informatica per le Aziende Digitali).

    Autenticazione: un singolo utente amministratore. Le operazioni di sola
    lettura su alimenti, LARN, ricette e smartfood sono pubbliche; le
    operazioni di scrittura su questi moduli e TUTTE le operazioni sui
    pazienti (dati clinici) richiedono il login (vedi /auth/login). La
    sessione e' mantenuta con un cookie firmato.
  version: "1.0.0"
servers:
  - url: /api
    description: Stesso host del front-end (modalita' consigliata)
...
security:
  - cookieAuth: []
paths:
...
  /ricette:
    get:
      tags: [Ricette]
      summary: Elenco paginato ricette (con filtri dieta/categoria/difficolta)
      security: []
      parameters:
        - $ref: "#/components/parameters/q"
        - $ref: "#/components/parameters/page"
        - name: categoria_ricetta
          in: query
          schema: { type: string }
        - name: difficolta
          in: query
          schema: { type: string, enum: [Facile, Media, Difficile] }
        - name: vegano
          in: query
          schema: { type: string, enum: ["0", "1"] }
        - name: vegetariano
          in: query
          schema: { type: string, enum: ["0", "1"] }
        - name: senza_glutine
          in: query
          schema: { type: string, enum: ["0", "1"] }
      responses:
        "200": { description: Elenco ricette }
    post:
      tags: [Ricette]
      summary: Crea una ricetta (senza ingredienti)
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/RicettaInput" } }
      responses:
        "201":
          description: Ricetta creata
          content:
            application/json: { schema: { $ref: "#/components/schemas/Ricetta" } }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
...
  /ricette/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [Ricette]
      summary: Dettaglio ricetta (testata)
      security: []
      responses:
        "200":
          description: Ricetta
          content:
            application/json: { schema: { $ref: "#/components/schemas/Ricetta" } }
        "404": { $ref: "#/components/responses/NonTrovato" }
    put:
      tags: [Ricette]
      summary: Aggiorna la testata di una ricetta
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/RicettaInput" } }
      responses:
        "200": { description: Ricetta aggiornata }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    delete:
      tags: [Ricette]
      summary: Elimina una ricetta (CASCADE sugli ingredienti)
      responses:
        "204": { description: Eliminata }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /ricette/{id}/valori-calcolati:
    parameters: [ { $ref: "#/components/parameters/idPath" } ]
    get:
      tags: [Ricette]
      summary: Valori nutrizionali per porzione calcolati "a volo" dagli ingredienti (non salva)
      security: []
      responses:
        "200": { description: Valori calcolati }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /ricette/{id}/ricalcola:
    parameters: [ { $ref: "#/components/parameters/idPath" } ]
    post:
      tags: [Ricette]
      summary: Ricalcola e SALVA i valori per porzione dagli ingredienti collegati
      responses:
        "200": { description: Ricetta aggiornata con i nuovi valori }
        "400": { description: Nessun ingrediente collegato }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
  /ricette/{id}/ingredienti:
    parameters: [ { $ref: "#/components/parameters/idPath" } ]
    get:
      tags: [Ricette]
      summary: Ingredienti della ricetta (con nome e kcal da alimenti)
      security: []
      responses:
        "200": { description: Elenco ingredienti }
        "404": { $ref: "#/components/responses/NonTrovato" }
    post:
      tags: [Ricette]
      summary: Aggiungi un ingrediente
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/RicettaIngredienteInput" } }
      responses:
        "201": { description: Ingrediente aggiunto }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
  /ricette/{id}/ingredienti/{ing_id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: ing_id
        in: path
        required: true
        schema: { type: integer }
    put:
      tags: [Ricette]
      summary: Modifica un ingrediente
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/RicettaIngredienteInput" } }
      responses:
        "200": { description: Ingrediente aggiornato }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    delete:
      tags: [Ricette]
      summary: Rimuovi un ingrediente
      responses:
        "204": { description: Rimosso }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }

  # ------------------------------------------------------------------
  # Smartfood
  # ------------------------------------------------------------------
...
  /smartfood:
    get:
      tags: [Smartfood]
      summary: Elenco righe (con filtri)
      security: []
      parameters:
        - $ref: "#/components/parameters/q"
        - name: frequenza
          in: query
          schema: { type: string }
        - name: gruppo
          in: query
          schema: { type: string }
      responses:
        "200": { description: Elenco righe }
    post:
      tags: [Smartfood]
      summary: Crea una riga
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/SmartfoodInput" } }
      responses:
        "201": { description: Riga creata }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
  /smartfood/{id}:
    parameters: [ { $ref: "#/components/parameters/idPath" } ]
    get:
      tags: [Smartfood]
      summary: Dettaglio riga
      security: []
      responses:
        "200": { description: Riga }
        "404": { $ref: "#/components/responses/NonTrovato" }
    put:
      tags: [Smartfood]
      summary: Aggiorna una riga
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/SmartfoodInput" } }
      responses:
        "200": { description: Riga aggiornata }
        "400": { $ref: "#/components/responses/DatiNonValidi" }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }
    delete:
      tags: [Smartfood]
      summary: Elimina una riga
      responses:
        "204": { description: Eliminata }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "404": { $ref: "#/components/responses/NonTrovato" }

  # ------------------------------------------------------------------
  # Pazienti
  # ------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie di sessione impostato da POST /auth/login.
  parameters:
    q:
      name: q
      in: query
      description: Ricerca libera
      schema: { type: string }
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    idPath:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    NonTrovato:
      description: Risorsa non trovata
      content:
        application/json: { schema: { $ref: "#/components/schemas/Errore" } }
    NonAutenticato:
      description: Login richiesto
      content:
        application/json: { schema: { $ref: "#/components/schemas/Errore" } }
    DatiNonValidi:
      description: Dati non validi
      content:
        application/json: { schema: { $ref: "#/components/schemas/ErroreValidazione" } }
  schemas:
    Errore:
      type: object
      properties:
        error: { type: string }
    ErroreValidazione:
      type: object
      properties:
        error: { type: string }
        details:
          type: array
          items: { type: string }
...
    RicettaInput:
      type: object
      required: [nome_ricetta]
      properties:
        nome_ricetta: { type: string }
        categoria_ricetta: { type: string, enum: [Antipasto, Primo, Secondo, Contorno, Dolce, Colazione] }
        difficolta: { type: string, enum: [Facile, Media, Difficile] }
        porzioni: { type: integer, minimum: 1, default: 4 }
        tempo_preparazione_min: { type: integer }
        vegano: { type: boolean }
        vegetariano: { type: boolean }
        senza_glutine: { type: boolean }
        procedimento: { type: string }
    Ricetta:
      allOf:
        - type: object
          properties:
            id: { type: integer }
            energia_kcal_porzione: { type: number, nullable: true }
            proteine_g_porzione: { type: number, nullable: true }
        - $ref: "#/components/schemas/RicettaInput"
    RicettaIngredienteInput:
      type: object
      required: [codice_alimento, quantita_g]
      properties:
        codice_alimento: { type: string }
        quantita_g: { type: number, exclusiveMinimum: 0 }
        ordine: { type: integer }
        note: { type: string }
    SmartfoodInput:
      type: object
      required: [alimento, gruppo_alimenti, frequenza_consumo]
      properties:
        alimento: { type: string }
        gruppo_alimenti: { type: string }
        frequenza_consumo: { type: string, enum: [giornaliera, settimanale, occasionale, evitare] }
        porzione_consigliata: { type: string }
        note: { type: string }
        ordine: { type: integer }
...